backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
reportlab==4.0.7
PyPDF2==3.0.1
pillow==10.1.0
email-validator==2.1.0
jinja2==3.1.2
aiosmtplib==3.0.1
cryptography>=42.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
